early-access version 1845

+++ src/video_core/texture_cache/types.h
@@ -133,8 +133,8 @@ struct BufferImageCopy {
 };
 
 struct BufferCopy {
-    u64 src_offset;
-    u64 dst_offset;
+    size_t src_offset;
+    size_t dst_offset;
     size_t size;
 };
